src/requirements.txt
Flask<2.0.0
Flask-Cors>=3.0.8
flask-restx==0.3.0
mysql-connector-python>=8.0.19
google-auth
requests